fixed more stuff, printAutopathing does nothing now
