+++ src/main/java/com/team1816/core/states/RobotState.java
@@ -1,12 +1,11 @@
 package com.team1816.core.states;
 
 import com.google.inject.Singleton;
-import com.team1816.lib.auto.Color;
-import com.team1816.lib.autopath.Autopath;
-import com.team1816.lib.subsystems.drive.SwerveDrive;
-import com.team1816.lib.util.visionUtil.VisionPoint;
 import com.team1816.core.configuration.Constants;
 import com.team1816.core.configuration.FieldConfig;
+import com.team1816.lib.auto.Color;
+import com.team1816.lib.subsystems.drive.SwerveDrive;
+import com.team1816.lib.util.visionUtil.VisionPoint;
 import com.team1816.season.auto.DynamicAutoScript2025;
 import com.team1816.season.subsystems.AlgaeCatcher;
 import com.team1816.season.subsystems.CoralArm;
@@ -15,7 +14,9 @@ import com.team1816.season.subsystems.Pneumatic;
 import edu.wpi.first.math.Matrix;
 import edu.wpi.first.math.VecBuilder;
 import edu.wpi.first.math.estimator.SwerveDrivePoseEstimator;
-import edu.wpi.first.math.geometry.*;
+import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.geometry.Rotation2d;
+import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
 import edu.wpi.first.math.kinematics.SwerveModulePosition;
 import edu.wpi.first.math.numbers.N1;
@@ -115,7 +116,7 @@ public class RobotState {
      */
 
     public boolean autopathing = false;
-    public boolean printAutopathing = false;  //Change this one to see the obstacle boundaries
+    public boolean printAutopathing = false;  //Change this one to see the obstacle boundaries //As of 2/8/2025, does nothing because of commented code in outputToSmartDashboard()
     public boolean printAutopathFieldTest = false;
     public Trajectory autopathTrajectory = null;
     public ArrayList<Trajectory> autopathTrajectoryPossibilities = new ArrayList<>();
@@ -241,49 +242,49 @@ public class RobotState {
     public synchronized void outputToSmartDashboard() {
         field.setRobotPose(fieldToVehicle);
 
-        if (printAutopathing) {
-            if (Autopath.fieldMap != null && Autopath.fieldMap.outputToSmartDashboardChanged) {
-                ArrayList<Pose2d> obstaclesExpanded = new ArrayList<>();
-
-                for (int i = 0; i < Autopath.fieldMap.getCurrentMap().getMapX(); i++) {
-                    for (int i2 = 0; i2 < Autopath.fieldMap.getCurrentMap().getMapY(); i2++) {
-                        if (Autopath.fieldMap.getCurrentMap().checkPixelHasObjectOrOffMap(i, i2)) {
-                            obstaclesExpanded.add(new Pose2d(new Translation2d(i / Autopath.mapResolution1DPerMeter, i2 / Autopath.mapResolution1DPerMeter), new Rotation2d()));
-                        }
-                    }
-                }
-
-                field.getObject("ExpandedObstacles").setPoses(obstaclesExpanded);
-
-                ArrayList<Pose2d> obstacles = new ArrayList<>();
-
-                for (int i = 0; i < Autopath.fieldMap.getCurrentMap().getMapX(); i++) {
-                    for (int i2 = 0; i2 < Autopath.fieldMap.getCurrentMap().getMapY(); i2++) {
-                        if (Autopath.fieldMap.getStableMapCheckPixelHasObjectOrOffMap(i, i2)) {
-                            obstacles.add(new Pose2d(new Translation2d(i / Autopath.mapResolution1DPerMeter, i2 /Autopath.mapResolution1DPerMeter), new Rotation2d()));
-                        }
-                    }
-                }
-
-                field.getObject("Obstacles").setPoses(obstacles);
-
-                Autopath.fieldMap.outputToSmartDashboardChanged = false;
-            }
-
-            if(autopathTrajectoryPossibilitiesChanged) {
-                for (int i = 0; i < autopathTrajectoryPossibilities.size(); i++) {
-                    if (autopathTrajectoryPossibilities.get(i) != null) {
-                        field.getObject("AutopathTrajectory: " + i).setTrajectory(autopathTrajectoryPossibilities.get(i));
-                    }
-                }
-                autopathMaxBranches = Math.max(autopathTrajectoryPossibilities.size(), autopathMaxBranches);
-                autopathTrajectoryPossibilitiesChanged = false;
-            }
-
-            field.getObject("StartCollisionPoints").setPoses(autopathCollisionStarts);
-            field.getObject("EndCollisionPoints").setPoses(autopathCollisionEnds);
-            field.getObject("AutopathWaypoints").setPoses(autopathWaypoints);
-        }
+//        if (printAutopathing) {
+//            if (Autopath.fieldMap != null && Autopath.fieldMap.outputToSmartDashboardChanged) {
+//                ArrayList<Pose2d> obstaclesExpanded = new ArrayList<>();
+//
+//                for (int i = 0; i < Autopath.fieldMap.getCurrentMap().getMapX(); i++) {
+//                    for (int i2 = 0; i2 < Autopath.fieldMap.getCurrentMap().getMapY(); i2++) {
+//                        if (Autopath.fieldMap.getCurrentMap().checkPixelHasObjectOrOffMap(i, i2)) {
+//                            obstaclesExpanded.add(new Pose2d(new Translation2d(i / Autopath.mapResolution1DPerMeter, i2 / Autopath.mapResolution1DPerMeter), new Rotation2d()));
+//                        }
+//                    }
+//                }
+//
+//                field.getObject("ExpandedObstacles").setPoses(obstaclesExpanded);
+//
+//                ArrayList<Pose2d> obstacles = new ArrayList<>();
+//
+//                for (int i = 0; i < Autopath.fieldMap.getCurrentMap().getMapX(); i++) {
+//                    for (int i2 = 0; i2 < Autopath.fieldMap.getCurrentMap().getMapY(); i2++) {
+//                        if (Autopath.fieldMap.getStableMapCheckPixelHasObjectOrOffMap(i, i2)) {
+//                            obstacles.add(new Pose2d(new Translation2d(i / Autopath.mapResolution1DPerMeter, i2 /Autopath.mapResolution1DPerMeter), new Rotation2d()));
+//                        }
+//                    }
+//                }
+//
+//                field.getObject("Obstacles").setPoses(obstacles);
+//
+//                Autopath.fieldMap.outputToSmartDashboardChanged = false;
+//            }
+//
+//            if(autopathTrajectoryPossibilitiesChanged) {
+//                for (int i = 0; i < autopathTrajectoryPossibilities.size(); i++) {
+//                    if (autopathTrajectoryPossibilities.get(i) != null) {
+//                        field.getObject("AutopathTrajectory: " + i).setTrajectory(autopathTrajectoryPossibilities.get(i));
+//                    }
+//                }
+//                autopathMaxBranches = Math.max(autopathTrajectoryPossibilities.size(), autopathMaxBranches);
+//                autopathTrajectoryPossibilitiesChanged = false;
+//            }
+//
+//            field.getObject("StartCollisionPoints").setPoses(autopathCollisionStarts);
+//            field.getObject("EndCollisionPoints").setPoses(autopathCollisionEnds);
+//            field.getObject("AutopathWaypoints").setPoses(autopathWaypoints);
+//        }
 
         if (autopathTrajectoryChanged && printAutopathing) {
             if(autopathTrajectory != null){
